requirements.txt
jupyterthemes == 0.20.0
plotly == 5.0.0
RISE == 5.7.1
wordcloud == 1.8.1
openpyxl
seaborn
matplotlib
japanize_matplotlib
polars
pyfpgrowth
networkx
pandas-profiling
dtale
sweetviz
matplotlib_venn
xgboost
pyarrow
fastparquet
catboost
# lightgbm
# gensim
# transformers[ja]
